добавил работу с сетью на экране каналов

+++ app/src/main/java/com/mironov/coursework/data/network/model/streams/StreamDto.kt
@@ -5,6 +5,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class StreamDto(
-    @SerialName("stream_id") val streamId: Long,
+    @SerialName("stream_id") val streamId: Int,
     @SerialName("name") val name: String,
 )
